+++ install-stubs/Orchid/Screens/User/UserEditScreen.php
@@ -8,10 +8,10 @@ use Orchid\Screen\Link;
 use Orchid\Screen\Layout;
 use Orchid\Screen\Screen;
 use Illuminate\Http\Request;
+use Orchid\Access\UserSwitch;
 use Orchid\Platform\Models\User;
 use Orchid\Support\Facades\Alert;
 use Orchid\Screen\Fields\Password;
-use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 use App\Orchid\Layouts\User\UserEditLayout;
 use App\Orchid\Layouts\User\UserRoleLayout;
@@ -31,6 +31,11 @@ class UserEditScreen extends Screen
      * @var string
      */
     public $description = 'All registered users';
+
+    /**
+     * @var string
+     */
+    public $permission = 'platform.systems.users';
 
     /**
      * Query data.
@@ -63,7 +68,7 @@ class UserEditScreen extends Screen
                 ->group([
                     Link::name(__('Login as user'))
                         ->icon('icon-login')
-                        ->method('switchUserStart'),
+                        ->method('loginAs'),
 
                     Link::name(__('Change Password'))
                         ->icon('icon-lock-open')
@@ -134,14 +139,14 @@ class UserEditScreen extends Screen
     }
 
     /**
-     * @param $id
+     * @param User $user
+     *
+     * @throws \Exception
      *
      * @return \Illuminate\Http\RedirectResponse
      */
-    public function remove($id)
+    public function remove(User $user)
     {
-        $user = User::findOrNew($id);
-
         $user->delete();
 
         Alert::info(__('User was removed'));
@@ -150,29 +155,15 @@ class UserEditScreen extends Screen
     }
 
     /**
-     * @param \Orchid\Platform\Models\User $user
+     * @param User $user
      *
      * @return \Illuminate\Http\RedirectResponse
      */
-    public function switchUserStart(User $user, Request $request)
+    public function loginAs(User $user)
     {
-        if (! session()->has('original_user')) {
-            session()->put('original_user', $request->user()->id);
-        }
-        Auth::login($user);
+        UserSwitch::loginAs($user);
 
-        return back();
-    }
-
-    /**
-     * @return \Illuminate\Http\RedirectResponse
-     */
-    public function switchUserStop()
-    {
-        $id = session()->pull('original_user');
-        Auth::loginUsingId($id);
-
-        return back();
+        return redirect()->route(config('platform.index'));
     }
 
     /**