+++ src/Platform/Dashboard.php
@@ -14,7 +14,7 @@ class Dashboard
     /**
      * ORCHID Version.
      */
-    public const VERSION = '5.0.0';
+    public const VERSION = '5.0.4';
 
     /**
      * The Dashboard configuration options.