+++ src/Platform/Http/Controllers/Auth/LoginController.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace Orchid\Platform\Http\Controllers\Auth;
 
 use Illuminate\View\View;
+use Orchid\Access\UserSwitch;
 use Illuminate\Cookie\CookieJar;
-use Illuminate\Config\Repository;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Validation\ValidationException;
 use Orchid\Platform\Http\Controllers\Controller;
@@ -32,7 +32,12 @@ class LoginController extends Controller
      */
     public function __construct()
     {
-        $this->middleware('guest', ['except' => 'logout']);
+        $this->middleware('guest', [
+                'except' => [
+                    'logout',
+                    'switchLogout',
+                ],
+            ]);
     }
 
     /**
@@ -46,11 +51,11 @@ class LoginController extends Controller
     /**
      * Where to redirect users after login / registration.
      *
-     * @return Repository|mixed
+     * @return string
      */
     public function redirectTo()
     {
-        return config('platform.prefix');
+        return route(config('platform.index'));
     }
 
     /**
@@ -75,5 +80,15 @@ class LoginController extends Controller
         $lockUser = $cookieJar->forget('lockUser');
 
         return redirect()->route('platform.login')->withCookie($lockUser);
+    }
+
+    /**
+     * @return \Illuminate\Http\RedirectResponse
+     */
+    public function switchLogout()
+    {
+        UserSwitch::logout();
+
+        return redirect()->route(config('platform.index'));
     }
 }